cleanup tests and init
